+++ cloudsql-postgresql-plugin/pom.xml
@@ -102,12 +102,7 @@
       <version>42.3.1</version>
       <scope>test</scope>
     </dependency>
-      <dependency>
-          <groupId>com.google.code.gson</groupId>
-          <artifactId>gson</artifactId>
-          <version>2.8.8</version>
-          <scope>test</scope>
-      </dependency>
+
   </dependencies>
 
   <build>
