+++ cloudsql-postgresql-plugin/src/e2e-test/java/io/cdap/plugin/cloudsqlpostgresql/stepsdesign/CloudSqlPostgreSql.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright © 2023 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package io.cdap.plugin.cloudsqlpostgresql.stepsdesign;
 
 import io.cdap.e2e.pages.actions.CdfPipelineRunAction;
@@ -5,7 +20,7 @@ import io.cdap.e2e.utils.BigQueryClient;
 import io.cdap.e2e.utils.CdfHelper;
 import io.cdap.e2e.utils.PluginPropertyUtils;
 import io.cdap.plugin.cloudsqlpostgresql.BQValidation;
-import io.cdap.plugin.cloudsqlpostgresql.CloudSqlPostgreSqlClient;
+import io.cdap.plugin.CloudSqlPostgreSqlClient;
 import io.cucumber.java.en.Then;
 import org.junit.Assert;
 import stepsdesign.BeforeActions;
@@ -15,14 +30,9 @@ import java.sql.SQLException;
 import java.text.ParseException;
 
 /**
- *  CLOUDSQLPOSTGRESQL Plugin related step design.
+ *  CLOUDSQL POSTGRESQL Step design.
  */
 public class CloudSqlPostgreSql implements CdfHelper {
-
-  @Then("Click on preview data for CloudSQLPostgreSQL sink")
-  public void clickOnPreviewDataForCloudSQLPostgreSQLSink() {
-    openSinkPluginPreviewData("CloudSQLPostgreSQL");
-  }
 
   @Then("Validate the values of records transferred to target table is equal to the values from source table")
   public void validateTheValuesOfRecordsTransferredToTargetTableIsEqualToTheValuesFromSourceTable() throws
