+++ cloudsql-postgresql-plugin/src/e2e-test/resources/pluginParameters.properties
@@ -1,8 +1,10 @@
 driverName=cloudsql-postgresql
 username=CLOUDSQL_POSTGRESQL_USERNAME
 password=CLOUDSQL_POSTGRESQL_PASSWORD
+connectionName=CLOUDSQL_POSTGRESQL_CONNECTION_NAME
 databaseName=test_automation_db
 schema=public
+URL=jdbc:postgresql://google/%s?cloudSqlInstance=%s&socketFactory=com.google.cloud.sql.postgres.SocketFactory&user=%s&password=%s
 
 datatypesColumns=( id varchar(100) primary key, col1 bpchar, col2 bpchar(10), col3 varchar, col4 varchar(3), \
   col5 bytea, col6 int2, col7 int4, col8 int8, col10 numeric(10, 4), col11 numeric(10), col12 float4, col13 float8, \
@@ -13,13 +15,11 @@ datatypesColumns=( id varchar(100) primary key, col1 bpchar, col2 bpchar(10), co
   col47 tsvector, col48 txid_snapshot, col49 uuid, col50 xml, col51 int4range, col52 int8range, col53 numrange, \
   col54 tsrange, col55 tstzrange, col56 daterange, col57 pg_lsn, col58 int4, col59 int2, col60 int8, col61 real, \
   col62 smallint, col63 serial, col64 smallserial, col65 double precision, col66 bigint, col67 bigserial, col68 boolean)
-
 datatypesColumnsList=( id, col1, col2, col3, col4, col5, col6 , col7 , col8 , col10, col11, col12, col13, col14, \
   col15, col16, col17, col18, col22, col23, col24, col25, col26, col27, col28, col29, col30, col31, col32, col33, \
   col34, col35, col36, col37, col38, col39, col40, col41, col42, col43, col44, col45, col46, col47, col48, col49, \
   col50, col51, col52, col53, col54, col55, col56, col57, col58, col59, col60, col61, col62, col63, col64, col65,\
   col66, col67, col68 )
-
 datatypesValues=VALUES ('User5', 'M', 'ABC...1234', 'B', 'ABC', decode('48656C6C6F20576F726C6421','hex'), 123, 123, \
   123456, 123.4567, 123456789, 123.456, 123.456, 100.26, 'Hello World!', 'User 5', 123.456, 100, \
   '2023-01-01 07:30:00.000', '2023-01-01 15:30:00.000', '02:00:00', '6 mons 02:30:00'::interval, \
@@ -35,7 +35,6 @@ datatypesValues=VALUES ('User5', 'M', 'ABC...1234', 'B', 'ABC', decode('48656C6C
   '["2010-01-01 20:00:00+05:30","2010-01-01 21:00:00+05:30")'::tstzrange, '[1992-03-21,1994-06-26)'::daterange, \
   '16/B374D848'::pg_lsn, 2, 2, 2, '1234.5679', '600', DEFAULT, DEFAULT, '61.823765812', '2500000000000', \
   DEFAULT, false);
-
 datatypesSchema=[{"key":"id","value":"string"},{"key":"col1","value":"string"},{"key":"col2","value":"string"},\
   {"key":"col3","value":"string"},{"key":"col4","value":"string"},{"key":"col5","value":"bytes"},\
   {"key":"col6","value":"int"},{"key":"col7","value":"int"},{"key":"col8","value":"long"},\
@@ -78,13 +77,12 @@ connectionTimeout=150
 numberOfSplits=2
 zeroValue=0
 splitByColumn=ID
-importQuery =  where $CONDITIONS
-connectionName=CONNECTION_NAME
+importQuery=where $CONDITIONS
 
 #bq properties
 projectId=cdf-athena
 dataset=test_automation
-bqOutputMultipleDatatypesSchema= [{"key":"col1","value":"bytes"},{"key":"col2","value":"string"},\
+bqOutputMultipleDatatypesSchema=[{"key":"col1","value":"bytes"},{"key":"col2","value":"string"},\
   {"key":"col3","value":"date"},{"key":"col4","value":"double"},{"key":"col5","value":"decimal"},\
   {"key":"col6","value":"timestamp"},{"key":"col7","value":"decimal"},{"key":"col8","value":"boolean"},\
   {"key":"col9","value":"long"},{"key":"col10","value":"time"}]
